Added RSQL for Specifications construction. Added paged list on UsersController.

+++ src/main/java/domain/responses/gets/lists/ResponseGetPagedList.java
@@ -2,6 +2,6 @@ package domain.responses.gets.lists;
 
 import java.util.List;
 
-public record ResponseGetPagedList<TResponseGetList>(int pageCount,
+public record ResponseGetPagedList<TResponseGetList>(long pageCount,
                                                      List<TResponseGetList> pageItems,
-                                                     int totalCount) { }
+                                                     long totalCount) { }
